#12 api call refactored and test passes
