+++ app/controllers/price.server.controller.js
@@ -1,12 +1,5 @@
-var request = require('request');
+var apiCall = require('./apiCall.server.controller.js');
 
-exports.request = function(req, res) {
-  request('https://openapi.etsy.com/v2/listings/active?api_key=96kg7j4073g5j4nc9tk5usgw', function(error, response, body) {
-    if (!error && response.statusCode === 200) {
-      var etsyData = JSON.parse(body);
-      res.json(etsyData);
-    } else {
-      console.log(error);
-    }
-  });
+exports.request = function(req,res){
+  apiCall(req,res);
 };
